Organizing Storybook stories and adding addons for Storybook.
