+++ README.md
@@ -1,35 +1,43 @@
 # reusable-components
-Reusable React components shared across the Library Simplified interfaces.
+Reusable React components shared across the Library Simplified interfaces. 
 
-<!-- [![npm version](https://badge.fury.io/js/simplified-circulation-web.svg)](https://badge.fury.io/js/simplified-circulation-web)
+[![npm version](https://badge.fury.io/js/library-simplified-reusable-components.svg)](https://badge.fury.io/js/library-simplified-reusable-components)
 
-[![Build Status](https://travis-ci.org/NYPL-Simplified/circulation-web.svg?branch=master)](https://travis-ci.org/NYPL-Simplified/circulation-web) -->
+[![Build Status](https://travis-ci.com/NYPL-Simplified/reusable-components.svg?branch=development)](https://travis-ci.com/NYPL-Simplified/reusable-components)
 
 ## Library Simplified Documentation
 
 To see screenshots, read in-depth documentation, and find out more about the project, check out the [Confluence](https://confluence.nypl.org/display/SIM/) site hosted by The New York Public Library.
 
-#### Use npm version
+To find documentation for the components in this repo (a work in progress), find it on our [Storybook documentation page](https://nypl-simplified.github.io/reusable-components).
 
-If you're working on one of the administrative interface and want to import reusable components from this repository:
+## npm package
 
-Instructions TBA!
+This package is [published to npm](https://www.npmjs.com/package/library-simplified-reusable-components).
 
-Webpack will take care of compiling and updating any new changes made locally for development.
+The `library-simplified-reusable-components` components is currently used in the Circulation Admin and the Registry Admin imported through npm. If you want to import the React components into your app, simply run `npm install --save library-simplified-reusable-components`.
 
-## Publishing
+### Publishing
 
-<!-- This package is [published to npm](https://www.npmjs.com/package/simplified-reusable-components). -->
-
-To publish a new version, you need to create an npm account and be a collaborator on the package. Then you can run `npm publish` from your local copy of the repository.
+To publish a new version, you need to have an npm account and be a collaborator on the package. Once a new feature or update has been merged into `master`, then you can run `npm publish` locally to publish a new version. Make sure it is on the `master` branch and that `package.json` is updated accordingly.
 
 ## Tests
 
-Like the codebase, all the unit tests are written in Typescript. Tests are written for all React components as well as redux and utility functions, and all can be found in their respective `__tests__` folders.
+Like the codebase, all the unit tests are written in Typescript. Tests are written for all React components and can be found in the `__tests__` folder.
 
 To run the tests, perform `npm test`.
 
 We use Travis CI for continuous integration. Any pull requests submitted must have tests and those tests must pass on Travis CI.
+
+## Storybook
+
+We are using [Storybook](https://storybook.js.org/) to document the React components found in this package.
+
+To run Storybook locally, run `npm run storybook` and the Storybook server will start and go to localhost:6006 automatically. Any changes to the stories files under `/stories` will hot reload.
+
+### Github Pages
+
+You can build and deploy a static version of the Storybook documentation _without_ having to commit and push changes to master. This is not recommended as the documentation can then be ahead of the `master` branch. Once a feature branch or any updates are merged into `master`, it is recommended to _then_ deploy the static build of the Storybook documentation. This can be done by running `npm run deploy-storybook`. If you have access to the Github repo, the `gh-pages` branch will automatically be updated and you can view updates on [Github](https://nypl-simplified.github.io/reusable-components/).
 
 ## License
