+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "library-simplified-reusable-components",
-  "version": "1.1.2",
+  "version": "1.2.0",
   "description": "Reusable React components for Library Simplified interfaces",
   "repository": {
     "type": "git",
@@ -16,7 +16,7 @@
     "test-file": "npm run lint && tsc && cp -r src/stylesheets lib && mocha --require lib/testHelper.js",
     "dev": "webpack --progress --display-modules --config webpack.dev.js",
     "prod": "webpack --progress --display-modules --config webpack.prod.js",
-    "storybook": "start-storybook -p 6006",
+    "storybook": "start-storybook -p 6006 -s ./dist",
     "build-storybook": "build-storybook",
     "deploy-storybook": "storybook-to-ghpages"
   },
@@ -27,11 +27,15 @@
   },
   "devDependencies": {
     "@babel/core": "^7.4.0",
+    "@storybook/addon-a11y": "^5.0.6",
     "@storybook/addon-actions": "^5.0.6",
+    "@storybook/addon-knobs": "^5.0.6",
     "@storybook/addon-links": "^5.0.6",
+    "@storybook/addon-storysource": "^5.0.6",
     "@storybook/addons": "^5.0.6",
     "@storybook/react": "^5.0.6",
     "@storybook/storybook-deployer": "^2.8.1",
+    "@storybook/theming": "^5.0.6",
     "babel-loader": "^8.0.5",
     "bootstrap": "^3.3.6",
     "chai": "^3.5.0",
