+++ src/components/Header.tsx
@@ -24,10 +24,12 @@ export default class Header extends React.Component<HeaderProps, void> {
             <img className="navbar-brand img-rounded" alt={this.props.alt || ""} src={src} />
           }
         </div>
-        <a href={this.props.logOut} className="btn navbar-btn">
-          <span>Log Out</span>
-          <LogoutIcon />
-        </a>
+        { this.props.logOut && (
+          <a href={this.props.logOut} className="btn navbar-btn">
+            <span>Log Out</span>
+            <LogoutIcon />
+          </a>)
+        }
       </nav>
     );
   }
